dégage les Long pour les int + début controller

+++ src/main/java/TopChefRPG/Repository/IngredientRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IngredientRepository extends JpaRepository<Ingredient,Long>{
+public interface IngredientRepository extends JpaRepository<Ingredient,Integer>{
 }
